+++ libs/request/src/RequestType.ts
@@ -1,6 +1,0 @@
-/**
- * The content type of the request.
- *
- * @public
- */
-export { RequestType } from '@svta/cml-utils'