+++ libs/request/src/ResponseType.ts
@@ -1,0 +1,6 @@
+/**
+ * The response type of the request.
+ *
+ * @public
+ */
+export { ResponseType } from '@svta/cml-utils'