+++ libs/utils/src/ResponseType.ts
@@ -1,22 +1,22 @@
 import type { ValueOf } from './ValueOf.ts'
 
 /**
- * The content type of the request.
- *
+ * The response type of the request.
  *
  * @enum
  *
  * @public
  */
-export const RequestType = {
+export const ResponseType = {
 	TEXT: 'text' as const,
 	JSON: 'json' as const,
 	BLOB: 'blob' as const,
 	ARRAY_BUFFER: 'arrayBuffer' as const,
 	DOCUMENT: 'document' as const,
+	STREAM: 'stream' as const,
 } as const
 
 /**
  * @public
  */
-export type RequestType = ValueOf<typeof RequestType>;
+export type ResponseType = ValueOf<typeof ResponseType>;